+++ src/materials/dielectric.h
@@ -7,10 +7,10 @@
 
 class Dielectric : public Material {
 public:
-	Dielectric(double refraction_index);
+	Dielectric(float refraction_index);
 
 	bool scatter(const Ray& r, const HitRecord& record, Color& attenuation_out, Ray& scattered_out) const override;
 
 private:
-	double _refraction_index;
+	float _refraction_index;
 };
